feat(project-two): implement LLM backend service using FastAPI
This commit initializes the backend service for Project 2, wrapping LLM capabilities into a RESTful API.

Key changes:
- Initialize FastAPI application with Uvicorn server configuration.
- Implement `/api/translate` endpoint with Pydantic validation.
- Create modular project structure (services, config, schemas).
- Refactor system prompts into a dedicated `PromptFactory` class.
- Integrate OpenAI/Aliyun SDK for model interaction.
- Add environment variable configuration support.
