--- llmcalling/project_two/prompt_factory.py
@@ -1,4 +1,0 @@
-translate_prompt = """
-"""
-
-system_prompt = {"translate": translate_prompt}
